# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForgeFit Hunter System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface UserProfile {
  id: number;
  userId: string;
  name: string;
  age: number;
  gender: string;
  height: number;
  currentWeight: number;
  targetWeight: number;
  goals: string[];
  experience: string;
  daysPerWeek: number;
  sessionLength: number;
  equipment: string[];
  diet: string;
  foodCulture: string;
  activityLevel: string;
  injuries: string;
  avoidFoods: string;
  updatedAt: string;
}

export interface UserProfileInput {
  name: string;
  age: number;
  gender: string;
  height: number;
  currentWeight: number;
  targetWeight: number;
  goals: string[];
  experience: string;
  daysPerWeek: number;
  sessionLength: number;
  equipment: string[];
  diet: string;
  foodCulture: string;
  activityLevel: string;
  injuries?: string;
  avoidFoods?: string;
}

export interface WorkoutLog {
  id: number;
  userId: string;
  date: string;
  day: string;
  exercises: string[];
  completed: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface WorkoutLogInput {
  day: string;
  exercises: string[];
  completed: boolean;
  notes?: string;
}

export interface WeightLog {
  id: number;
  userId: string;
  weight: number;
  date: string;
  createdAt: string;
}

export interface WeightLogInput {
  weight: number;
  date?: string;
}

export interface DashboardStats {
  weeklyWorkouts: number;
  weeklyGoal: number;
  streakDays: number;
  disciplineScore: number;
  totalWorkouts: number;
  /** @nullable */
  latestWeight: number | null;
  /** @nullable */
  targetWeight: number | null;
}

/**
 * Opaque session token — Bearer <sid>.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};