# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ForgeFit Hunter System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: openai
    description: OpenAI AI chat and coaching operations
  - name: profile
    description: User profile management
  - name: logs
    description: Workout log operations
  - name: weightLogs
    description: Weight tracking operations
  - name: dashboard
    description: Dashboard stats
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the current user's ForgeFit profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
    put:
      operationId: upsertProfile
      tags: [profile]
      summary: Create or update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /logs:
    get:
      operationId: listWorkoutLogs
      tags: [logs]
      summary: List workout logs for the current user
      responses:
        "200":
          description: List of workout logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutLog"
    post:
      operationId: createWorkoutLog
      tags: [logs]
      summary: Log a completed workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutLogInput"
      responses:
        "201":
          description: Created log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutLog"
  /logs/{id}:
    delete:
      operationId: deleteWorkoutLog
      tags: [logs]
      summary: Delete a workout log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /weight-logs:
    get:
      operationId: listWeightLogs
      tags: [weightLogs]
      summary: List weight logs for the current user
      responses:
        "200":
          description: List of weight logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeightLog"
    post:
      operationId: createWeightLog
      tags: [weightLogs]
      summary: Add a weight entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeightLogInput"
      responses:
        "201":
          description: Created weight log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeightLog"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
          format: email
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    OpenaiConversation:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
    OpenaiMessage:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    OpenaiConversationInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
    OpenaiMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    OpenaiConversationWithMessages:
      type: object
      required: [id, title, createdAt, messages]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
    OpenaiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    UserProfile:
      type: object
      required: [id, userId, name, age, gender, height, currentWeight, targetWeight, goals, experience, daysPerWeek, sessionLength, equipment, diet, foodCulture, activityLevel, injuries, avoidFoods, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        age:
          type: integer
        gender:
          type: string
        height:
          type: number
        currentWeight:
          type: number
        targetWeight:
          type: number
        goals:
          type: array
          items:
            type: string
        experience:
          type: string
        daysPerWeek:
          type: integer
        sessionLength:
          type: integer
        equipment:
          type: array
          items:
            type: string
        diet:
          type: string
        foodCulture:
          type: string
        activityLevel:
          type: string
        injuries:
          type: string
        avoidFoods:
          type: string
        updatedAt:
          type: string
          format: date-time
    UserProfileInput:
      type: object
      required: [name, age, gender, height, currentWeight, targetWeight, goals, experience, daysPerWeek, sessionLength, equipment, diet, foodCulture, activityLevel]
      properties:
        name:
          type: string
        age:
          type: integer
        gender:
          type: string
        height:
          type: number
        currentWeight:
          type: number
        targetWeight:
          type: number
        goals:
          type: array
          items:
            type: string
        experience:
          type: string
        daysPerWeek:
          type: integer
        sessionLength:
          type: integer
        equipment:
          type: array
          items:
            type: string
        diet:
          type: string
        foodCulture:
          type: string
        activityLevel:
          type: string
        injuries:
          type: string
        avoidFoods:
          type: string

    WorkoutLog:
      type: object
      required: [id, userId, date, day, exercises, completed, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        date:
          type: string
          format: date-time
        day:
          type: string
        exercises:
          type: array
          items:
            type: string
        completed:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    WorkoutLogInput:
      type: object
      required: [day, exercises, completed]
      properties:
        day:
          type: string
        exercises:
          type: array
          items:
            type: string
        completed:
          type: boolean
        notes:
          type: string

    WeightLog:
      type: object
      required: [id, userId, weight, date, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        weight:
          type: number
        date:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    WeightLogInput:
      type: object
      required: [weight]
      properties:
        weight:
          type: number
        date:
          type: string
          format: date-time

    DashboardStats:
      type: object
      required: [weeklyWorkouts, weeklyGoal, streakDays, disciplineScore, totalWorkouts, latestWeight, targetWeight]
      properties:
        weeklyWorkouts:
          type: integer
        weeklyGoal:
          type: integer
        streakDays:
          type: integer
        disciplineScore:
          type: number
        totalWorkouts:
          type: integer
        latestWeight:
          type: ["number", "null"]
        targetWeight:
          type: ["number", "null"]