# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForgeFit Hunter System API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  weeklyWorkouts: number;
  weeklyGoal: number;
  streakDays: number;
  disciplineScore: number;
  totalWorkouts: number;
  /** @nullable */
  latestWeight: number | null;
  /** @nullable */
  targetWeight: number | null;
}