# lib/api-zod/src/generated/types/logoutSuccess.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForgeFit Hunter System API
 * OpenAPI spec version: 0.1.0
 */

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;