# lib/api-zod/src/generated/types/userProfileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForgeFit Hunter System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileInput {
  name: string;
  age: number;
  gender: string;
  height: number;
  currentWeight: number;
  targetWeight: number;
  goals: string[];
  experience: string;
  daysPerWeek: number;
  sessionLength: number;
  equipment: string[];
  diet: string;
  foodCulture: string;
  activityLevel: string;
  injuries?: string;
  avoidFoods?: string;
}