# lib/api-zod/src/generated/types/weightLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForgeFit Hunter System API
 * OpenAPI spec version: 0.1.0
 */

export interface WeightLog {
  id: number;
  userId: string;
  weight: number;
  date: Date;
  createdAt: Date;
}